requirements.txt
customtkinter
CTkMessageBox
pyserial
pywinstyles
Pillow